code format

+++ DataEditorX/Common/MyBitmap.cs
@@ -2,10 +2,10 @@
  * date :2014-02-07
  * desc :图像处理，裁剪，缩放，保存
  */
-using System.IO;
 using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Drawing.Imaging;
+using System.IO;
 
 namespace DataEditorX.Common
 {
@@ -35,7 +35,7 @@ namespace DataEditorX.Common
         /// <returns>处理好的图像</returns>
         public static Bitmap Zoom(Bitmap sourceBitmap, int newWidth, int newHeight)
         {
-            if ( sourceBitmap != null )
+            if (sourceBitmap != null)
             {
                 Bitmap b = new Bitmap(newWidth, newHeight);
                 Graphics graphics = Graphics.FromImage(b);
@@ -79,17 +79,17 @@ namespace DataEditorX.Common
         /// <returns>处理好的图像</returns>
         public static Bitmap Cut(Bitmap sourceBitmap, int StartX, int StartY, int cutWidth, int cutHeight)
         {
-            if ( sourceBitmap != null )
+            if (sourceBitmap != null)
             {
                 int w = sourceBitmap.Width;
                 int h = sourceBitmap.Height;
                 //裁剪的区域宽度调整
-                if ( ( StartX + cutWidth ) > w )
+                if ((StartX + cutWidth) > w)
                 {
                     cutWidth = w - StartX;
                 }
                 //裁剪的区域高度调整
-                if ( ( StartY + cutHeight ) > h )
+                if ((StartY + cutHeight) > h)
                 {
                     cutHeight = h - StartY;
                 }
@@ -121,12 +121,12 @@ namespace DataEditorX.Common
         /// <param name="filename">保存路径</param>
         /// <param name="quality">质量</param>
         /// <returns>是否保存成功</returns>
-        public static bool SaveAsJPEG(Bitmap bitmap, string filename, int quality=90)
+        public static bool SaveAsJPEG(Bitmap bitmap, string filename, int quality = 90)
         {
-            if ( bitmap != null )
+            if (bitmap != null)
             {
-            	string path=Path.GetDirectoryName(filename);
-            	if(!Directory.Exists(path))//创建文件夹
+                string path=Path.GetDirectoryName(filename);
+                if (!Directory.Exists(path))//创建文件夹
                 {
                     Directory.CreateDirectory(path);
                 }
@@ -138,9 +138,9 @@ namespace DataEditorX.Common
 
                 ImageCodecInfo[] codecs = ImageCodecInfo.GetImageEncoders();
                 ImageCodecInfo ici = null;
-                foreach ( ImageCodecInfo codec in codecs )
+                foreach (ImageCodecInfo codec in codecs)
                 {
-                    if ( codec.MimeType.IndexOf("jpeg") > -1 )
+                    if (codec.MimeType.IndexOf("jpeg") > -1)
                     {
                         ici = codec;
                         break;
@@ -152,7 +152,7 @@ namespace DataEditorX.Common
                 }
 
                 EncoderParameters encoderParams = new EncoderParameters();
-                encoderParams.Param[0] = new EncoderParameter(Encoder.Quality, (long)quality);
+                encoderParams.Param[0] = new EncoderParameter(Encoder.Quality, quality);
                 if (ici != null)
                 {
                     bitmap.Save(filename, ici, encoderParams);
